Refactor plugin to use module
