+++ test/plugin/test_filter_flowcounter_simple.rb
@@ -32,14 +32,14 @@ class FlowCounterSimpleFilterTest < Test::Unit::TestCase
   private
 
   def filter(d,  msgs)
-    stub(d.instance.output).start
-    stub(d.instance.output).shutdown
+    stub(d.instance).start
+    stub(d.instance).shutdown
     d.run {
       msgs.each {|msg|
         d.filter(msg, @time)
       }
     }
-    out = capture_log(d.instance.output.log) do
+    out = capture_log(d.instance.log) do
       d.instance.flush_emit(0)
     end
     filtered = d.filtered_as_array
